refactor: 타입 구조 개선 및 중복 제거
- types/grade.ts 생성하여 학점/어학성적 타입 분리
- types/slot.ts에서 Slot 타입 재사용 (extends, 중복 제거)
- types/application.ts 간소화 (지원서 관련 타입만 유지)
- 모든 grade 타입 import를 @/types/grade로 변경하여 명확성 향상
